Move get_gcloud_storage_args tests to test_shim_util
